lersosa-python/requirements.txt
fastapi~=0.115.12
grpcio~=1.72.0rc1
starlette~=0.46.2
nacos-sdk-python~=2.0.6
pydantic~=2.11.4
uvicorn~=0.34.2
grpcio-tools~=1.72.0rc1
python-multipart~=0.0.20
python-dotenv~=1.1.0

numpy~=2.2.5
matplotlib~=3.10.3
pillow~=11.2.1
tqdm~=4.67.1
six~=1.17.0
protobuf~=6.30.0
einops~=0.8.1
einops-exts~=0.0.4
requests~=2.32.3
websockets~=15.0.1
redis~=6.1.0

python-jose~=3.4.0
circuitbreaker~=2.1.3
torch
torchvision
